Rename PAX files to be consistent
